Merge azd env file with local env file instead of overwriting
The local env file will be used to source app env vars when running azd locally (instead of sourcing from AZDO/GitHub when running in a pipeline), but it would be annoying if your local env file keeps getting overwritten every time you run azd. So now the two are merged and the output is used in the Docker build.