Remove managed certificate from the infra scripts
There are various chicken and egg issues with automating provisioning of the managed cert for container apps. There are workarounds, but they are all so convoluted as to make it more hassle than its worth when a manual solution is available. Not ideal, but on balance the best solution for now.